implemented new logic, validation, liquibase transaction, circuit breaker pattern, etc.

+++ src/main/java/by/ares/orderservice/dto/response/OrderDto.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class OrderDto {
     private Long id;
     private Status status;
-    private Float totalPrice;
+    private Long totalPrice;
     private Boolean deleted;
     private UserDto userDto;
     private List<ItemDto> items;
